% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErrorModal.R
\name{ErrorModal}
\alias{ErrorModal}
\title{ErrorModal_UI}
\usage{
ErrorModal(input, output, session, title = "Error", subtitle = NULL,
  message = "Sorry, an error has occured", try_this)
}
\arguments{
\item{title}{Title for the Error box, defaults to "Error", but could be custom message}

\item{subtitle}{Optional subtitle to describe the error}

\item{message}{Ideally most likely cause for error. Example would be}

\item{try_this}{User specified function to run. This will be completed if successful or on error the modal box will be shown}
}
\description{
This shiny module UI runs a specified function and alerts tha the function was completed or shows an error with message parameter for most likely cause.
}
\examples{
callModule(ErrorModal,"TryThis_01",title="Error",subtitle=NULL,message='Sorry, invalid column selection',try_this=column_select(input$col))
}
\keyword{ErrorModal}
\keyword{ModalDialog}
\keyword{shinydashboard}
\keyword{tryCatch}
