% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErrorModal.R
\name{ErrorModal_UI}
\alias{ErrorModal_UI}
\title{ErrorModal_UI}
\usage{
ErrorModal_UI(id, label = "Go")
}
\arguments{
\item{id}{Unique Identifier for the UI in the shiny app UI}

\item{title}{InfoBox title, Default is "My InfoModalBox}

\item{value}{Value to be displayed, typically a summary value. Default is 100}

\item{color}{Color of InfoBox, Default is "green" For options see shiny InfoBox documentation}

\item{color}{Icon to display, it is a call to shiny's icon() function. Default is "thumbs-up"}
}
\description{
This shiny module UI runs a specified function and alerts tha the function was completed or shows an error with message parameter for most likely cause.
}
\examples{
ErrorModal_UI("RobustDo")
}
\keyword{InfoBox}
\keyword{ModalDialog}
\keyword{shinydashboard}
