% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfoBoxModal.R
\name{InfoBoxModal}
\alias{InfoBoxModal}
\title{InfoBoxModal}
\usage{
InfoBoxModal(input, output, session, title = "My InfoModalBox Modal Title",
  size = "l", elements)
}
\arguments{
\item{title}{Modal Dialog title}

\item{size}{Modal Dialog title}

\item{elements}{A tagList of UI elements to be shown in the modal box}
}
\description{
This shiny module server function registers a click event on the InfoBox and pulls up a shinyDialog Box. The function accepts whatever tagList of UI elements you wish to display in the pop up.
}
\examples{
callModule(InfoBoxModal,"InfoBox",tagList(fluidRow(column(6,h3('Header for Stuff for Modal')),column(6,h3('More Stuff')))))
}
\keyword{InfoBox}
\keyword{ModalDialog}
\keyword{shinydashboard}
