% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfoBoxModal.R
\name{InfoBoxModal_UI}
\alias{InfoBoxModal_UI}
\title{InfoBoxModal_UI}
\usage{
InfoBoxModal_UI(id, title = "My InfoModalBox", color = "green",
  value = 100, icon = "thumbs-up")
}
\arguments{
\item{id}{Unique Identifier for the UI in the shiny app UI}

\item{title}{InfoBox title, Default is "My InfoModalBox}

\item{color}{Color of InfoBox, Default is "green" For options see shiny InfoBox documentation}

\item{value}{Value to be displayed, typically a summary value. Default is 100}

\item{color}{Icon to display, it is a call to shiny's icon() function. Default is "thumbs-up"}
}
\description{
This shiny module UI creates an InfoBox that can be clicked on to bring up a modal dialof with expanded details of the summary in the InfoBox.
}
\examples{
InfoBoxModal_UI("InfoBox")
}
\keyword{InfoBox}
\keyword{ModalDialog}
\keyword{shinydashboard}
